Range bug fix
Issue was (I think) with jquery comparing the upper and lower ranges as
objects/Strings or something instead of as the numbers they are, so... I
just created a couple variables, parse the int values into those and
then compared those.
____________
PS
This is the same update as the HowTo branch (5/17/2016), accidently
pushed it to that one as well, both are the same at this point.